# api-docs/openapi/api-spec.yml
openapi: 3.0.0
info:
  title: HumanAIOS Canonical API
  version: 1.0.0
  description: Canonical NestJS API surface implemented under /apps/api.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
...
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a user and issue an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /auth/verify:
    post:
      tags: [Authentication]
      summary: Verify the current bearer token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/AuthenticatedUser'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
        org_name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, operator, viewer]
    AuthenticatedUser:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            user_id:
              type: string
...